socratic-dialogue-mvp/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
redis==6.4.0
pytest==8.4.1
httpx==0.28.1
python-dotenv==1.1.0
